add seed of continents and try many to many relationship - not working yet

+++ WildAnimalsAPI/Models/DTO/CreateContinentDto.cs
@@ -3,8 +3,8 @@
     public class CreateContinentDto
     {
         public string Name { get; set; }
-        public int Population { get; set; }
-        public int SurfaceKm2 { get; set; }
+        public long Population { get; set; }
+        public long SurfaceKm2 { get; set; }
 
         public List<AnimalContinent>? AnimalContinentMappings { get; set; }
     }
